moved some input handlers into their own files
