+++ demo/dialog/demo_dialog.c
@@ -38,7 +38,7 @@ void demo_init_demo_dialog(eg_app *app)
 
 void demo_open_demo_dialog(eg_app *app)
 {
-    eg_push_input_handler(app, demo_dialog_input_handler);
+    eg_push_input_handler(app, common_dialog_input_handler);
 
     // Reset the dialog.
     demo_dialog.panel = 0;
