+++ demo/dialog/info_dialog.c
@@ -43,7 +43,7 @@ void demo_init_info_dialog(eg_app *app)
 
 void demo_open_info_dialog(eg_app *app)
 {
-    eg_push_input_handler(app, demo_dialog_input_handler);
+    eg_push_input_handler(app, common_dialog_input_handler);
 
     // Reset the dialog.
     info_dialog.panel = 0;
