+++ demo/dialog/sign_dialog.c
@@ -40,7 +40,7 @@ void demo_init_sign_dialog(eg_app *app)
 
 void demo_open_sign_dialog(eg_app *app)
 {
-    eg_push_input_handler(app, demo_dialog_input_handler);
+    eg_push_input_handler(app, common_dialog_input_handler);
 
     // Reset the dialog.
     sign_dialog.panel = 0;
