+++ demo/input/input.c
@@ -16,7 +16,7 @@ void demo_init_input(eg_app *app)
     app->input = &(input_handlers[0]);
 }
 
-void demo_dialog_input_handler(eg_app *app)
+void common_dialog_input_handler(eg_app *app)
 {
     eg_dialog *d = app->dialogs[app->dialog_count - 1];
 
@@ -39,168 +39,6 @@ void demo_dialog_input_handler(eg_app *app)
         {
             d->advance(app, d);
         }
-    }
-}
-
-void fish_menu_input_handler(eg_app *app)
-{
-    if (eg_consume_input(app, EG_KEYCODE_X))
-    {
-        app->menu_count--;
-        eg_pop_input_handler(app);
-        return;
-    }
-
-    if (eg_consume_input(app, EG_KEYCODE_UP))
-    {
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        if (m->cursor.y > 0)
-        {
-            m->cursor.y--;
-        }
-    }
-
-    if (eg_consume_input(app, EG_KEYCODE_DOWN))
-    {
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        if (m->cursor.y < 1)
-        {
-            m->cursor.y++;
-        }
-    }
-
-    // menu item selection
-    if (eg_consume_input(app, EG_KEYCODE_Z))
-    {
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        // Determine which menu item has been selected.
-        // Layout:
-        // items[0]
-        // items[1]
-        eg_menu_item *item = m->items[m->cursor.y];
-
-        item->callback(app, m);
-    }
-}
-
-void info_menu_input_handler(eg_app *app)
-{
-    if (eg_consume_input(app, EG_KEYCODE_X))
-    {
-        app->menu_count--;
-        eg_pop_input_handler(app);
-
-        if (app->dialog_count > 0)
-        {
-            eg_dialog *d = app->dialogs[app->dialog_count - 1];
-            d->result = 0;
-            d->advance(app, d);
-        }
-
-        return;
-    }
-
-    if (eg_consume_input(app, EG_KEYCODE_UP))
-    {
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        if (m->cursor.y > 0)
-        {
-            m->cursor.y--;
-        }
-    }
-
-    if (eg_consume_input(app, EG_KEYCODE_DOWN))
-    {
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        if (m->cursor.y < 1)
-        {
-            m->cursor.y++;
-        }
-    }
-
-    // menu item selection
-    if (eg_consume_input(app, EG_KEYCODE_Z))
-    {
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        // Determine which menu item has been selected.
-        // Layout:
-        // items[0]
-        // items[1]
-        eg_menu_item *item = m->items[m->cursor.y];
-
-        item->callback(app, m);
-    }
-}
-
-void pause_menu_input_handler(eg_app *app)
-{
-    if (eg_consume_input(app, EG_KEYCODE_X) || eg_consume_input(app, EG_KEYCODE_Q))
-    {
-        app->menu_count--;
-        app->pause = 0;
-        eg_pop_input_handler(app);
-        return;
-    }
-
-    if (eg_consume_input(app, EG_KEYCODE_UP))
-    {
-        // Get the menu at the top of the stack.
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        if (m->cursor.y > 0)
-        {
-            m->cursor.y--;
-        }
-    }
-
-    if (eg_consume_input(app, EG_KEYCODE_DOWN))
-    {
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        if (m->cursor.y < 1)
-        {
-            m->cursor.y++;
-        }
-    }
-
-    if (eg_consume_input(app, EG_KEYCODE_LEFT))
-    {
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        if (m->cursor.x > 0)
-        {
-            m->cursor.x--;
-        }
-    }
-
-    if (eg_consume_input(app, EG_KEYCODE_RIGHT))
-    {
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        if (m->cursor.x < 1)
-        {
-            m->cursor.x++;
-        }
-    }
-
-    // menu item selection
-    if (eg_consume_input(app, EG_KEYCODE_Z))
-    {
-        eg_menu *m = app->menus[app->menu_count - 1];
-
-        // Determine which menu item has been selected.
-        // Layout:
-        // items[0]  items[2]
-        // items[1]  items[3]
-        eg_menu_item *item = m->items[m->cursor.y + 2 * m->cursor.x];
-
-        item->callback(app, NULL);
     }
 }
