+++ demo/input/input.h
@@ -13,7 +13,7 @@ void root_input_handler(eg_app *);
 void fish_menu_input_handler(eg_app *);
 void info_menu_input_handler(eg_app *);
 void pause_menu_input_handler(eg_app *);
-void demo_dialog_input_handler(eg_app *);
+void common_dialog_input_handler(eg_app *);
 void debug_menu_input_handler(eg_app *);
 void scenes_menu_input_handler(eg_app *);
 void input_menu_input_handler(eg_app *);
